+++ src/main/java/tech/autodirect/api/entities/EntOffer.java
@@ -16,17 +16,16 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-import java.math.BigDecimal;
 import java.sql.SQLException;
 import java.util.Map;
 
 public class EntOffer {
     private int offerId;
     private int carId;
-    private BigDecimal loanAmount;
-    private BigDecimal capitalSum;
-    private BigDecimal interestSum;
-    private BigDecimal totalSum;
+    private double loanAmount;
+    private double capitalSum;
+    private double interestSum;
+    private double totalSum;
     private double interestRate;
     private double termMo;
     private String installments;
@@ -40,10 +39,10 @@ public class EntOffer {
     public void loadFromList(Map<String, Object> entry) throws SQLException {
         offerId = (int) entry.get("offer_id");
         carId = (int) entry.get("car_id");
-        loanAmount = (BigDecimal) entry.get("loan_amount");
-        capitalSum = (BigDecimal) entry.get("capital_sum");
-        interestSum = (BigDecimal) entry.get("interest_sum");
-        totalSum = (BigDecimal) entry.get("total_sum");
+        loanAmount = (double) entry.get("loan_amount");
+        capitalSum = (double) entry.get("capital_sum");
+        interestSum = (double) entry.get("interest_sum");
+        totalSum = (double) entry.get("total_sum");
         interestRate = (double) entry.get("interest_rate");
         termMo = (double) entry.get("term_mo");
         installments = (String) entry.get("installments");
@@ -58,19 +57,19 @@ public class EntOffer {
         return carId;
     }
 
-    public BigDecimal getLoanAmount() {
+    public double getLoanAmount() {
         return loanAmount;
     }
 
-    public BigDecimal getCapitalSum() {
+    public double getCapitalSum() {
         return capitalSum;
     }
 
-    public BigDecimal getInterestSum() {
+    public double getInterestSum() {
         return interestSum;
     }
 
-    public BigDecimal getTotalSum() {
+    public double getTotalSum() {
         return totalSum;
     }
