+++ src/main/java/tech/autodirect/api/entities/EntUser.java
@@ -16,15 +16,14 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-import java.math.BigDecimal;
 import java.sql.SQLException;
 import java.util.Map;
 
 public class EntUser {
     private String userId;
     private int creditScore;
-    private BigDecimal downPayment;
-    private BigDecimal budgetMo;
+    private double downPayment;
+    private double budgetMo;
     private String offersTable;
 
     /**
@@ -35,8 +34,8 @@ public class EntUser {
     public void loadFromList(Map<String, Object> entry) throws SQLException {
         userId = (String) entry.get("user_id");
         creditScore = (int) entry.get("credit_score");
-        downPayment = (BigDecimal) entry.get("down_payment");
-        budgetMo = (BigDecimal) entry.get("budget_mo");
+        downPayment = (double) entry.get("down_payment");
+        budgetMo = (double) entry.get("budget_mo");
         offersTable = (String) entry.get("offers_table");
     }
 
@@ -48,11 +47,11 @@ public class EntUser {
         return creditScore;
     }
 
-    public BigDecimal getDownPayment() {
+    public double getDownPayment() {
         return downPayment;
     }
 
-    public BigDecimal getBudgetMo() {
+    public double getBudgetMo() {
         return budgetMo;
     }
